% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regime.R
\name{regime}
\alias{regime}
\title{Spatial Regime Model}
\usage{
regime(
  restrict,
  group,
  data,
  var_int = FALSE,
  lw_lst = NULL,
  error = TRUE,
  robust = FALSE
)
}
\arguments{
\item{restrict}{\command{lm}-style formula for the restricted regression.}

\item{group}{RHS formula depicting only the factor variable
used to identify regimes (i.e. no dependent variable needed).}

\item{data}{model data frame.}

\item{var_int}{a logical value indicating whether or not to include
group-specific intercepts in the restricted model.  Default is \code{FALSE}}

\item{lw_lst}{a list of \code{listw} objects corresponding to the levels of 
the factor used to define regimes.  The default value is \code{NULL}.}

\item{error}{a logical value indicating type of spatial regime model to use
when \code{!is.null(lw_lst)}.  A value of \code{TRUE} produces a spatial
error model, while a value of \code{FALSE} produces a spatial lag model.
#' @param robust a logical value indicating whether to use HC3 standard 
errors.}

\item{...}{additional arguments to be passed to the model-fitting routine.}
}
\value{

}
\description{
Estimates the stability of a given model across a set of
user-specified regimes.
}
\details{
\command{regime} compares the fit of a restricted or global model in which a
given set of coefficients are treated as fixed to the fit of an unrestricted
or local model in which the coefficients in question are allowed to vary
across a set of user-specified regimes.  Regime models are specified using
two formulas, the first of which depicts the global model, the second of
which depicts the variable used to specify regimes.  Spatial regime models
are estimated using the \command{errorsarlm} and \command{lagsarlm} routines
available as part of the \pkg{spreg} package.  The resulting test is
ultimately carried out using the \command{anova} command which, in the case
of non-spatial regime models, results in the standard F-test.  In contrast,
the use of the \command{anova} command in the context of likelihood-based
models such as \command{errorsarlm} and \command{lagsarlm} results in a
likelihood ratio test.  The null model can be adjusted to allow for
regime-specific intercepts.
}
\examples{
data(nat60_sf)
reg_mod60 <- regime(HR60 ~ RD60 + PS60 + MA60 + DV60 + UE60, 
     group = ~ SOUTH, data = nat60_sf)
reg_mod60
}
\references{
Anselin, L.  1988.
\emph{Spatial Econometrics: Methods and Models}.
Durdrecht: Kluwer Academic Publishers, pp. 123-124.

Chow, G.  1960.  "Tests of Equality Between Sets of Coefficients
in Two Linear Regressions."  \emph{Econometrica} 28:591-605.
}
\author{
Adam Slez \email{aslez@virginia.edu}
}
